feat!: add client-side routing with React Router 6
I used the new data APIs that were introduced in React Router 6.4 to
implement client-side routing. In doing so, the <App /> component is
replaced with <RouterProvider /> as the root component.

For now, rendering EITHER App or RouterProvider component works. The
components have been refactored with conditional statements to work with
either approach. However, I'll be focusing on the client-side routing
moving forward.

Also added a Tailwind plugin that adds fluid custom properties to the
root.

BREAKING CHANGE: render <Root /> instead of <App /> as the root
component.