feat: make scroll behavior more predictable
Previously, visiting a country page using a border country button
doesn't reset the scroll position, so the user is already in the middle
of the new page.

This commit adds the `<ScrollRestoration />` component from React
Router, which emulates the browser's scroll behavior when moving to a
different page. So when a user taps on a country button, the new page's
scroll position is at the top of the page. Navigating forward/backward
preserves the previous scroll position, which helps to retain context so
the user can more easily find their way through the site.