refactor: implementation for hiding clear button
Previously we're using the presence of the `q` prop to set the clear
button's `hidden` attribute. This means that it only gets updated when
the searchbox component gets rerendered, which is after the debounced
submit function is called. Moving this implementation logic to the CSS
removes the lag from when the user types something in the searchbox to
the clear button appearing/disappearing.