Enhance search bar on mobile, esp. iOS Safari
First, there's an annoying input zoom when focusing an input on iOS
Safari. To fix this, I added `maximum-scale=1.0` to the viewport meta
tag.

Again on iOS Safari, the search icon appears twice, so I removed the
browser default styling with `-webkit-appearance: none`. I also replaced
the default label on the enter key when the search input is focused from
"return" to "search".

Finally, I've updated the submit handler to close the keypad and prevent
the default behavior of reloading the page.